ml/requirements.txt
fastapi
uvicorn
requests
pydantic
SQLAlchemy
python-dotenv